requirements.txt
defusedxml
gitpython
rich
